Improve image file deletion validation
The current code for updating a user's profile image has been updated for better file deletion validation. It now checks if the image file exists in the file system before attempting to delete it, preventing possible errors due to non-existent files. This improves robustness in the code handling of profile images.

+++ website/views.py
@@ -232,7 +232,9 @@ def edit_personal_info(request, pk):
 
                 if 'profile_image' in request.FILES:
                     # user updating image so delete the old one from file system
-                    personal_info.profile_image.delete()
+                    if personal_info.profile_image and os.path.isfile(
+                            os.path.join(settings.MEDIA_ROOT, personal_info.profile_image.path)):
+                        personal_info.profile_image.delete()
 
                     # upload new image
                     profile_image = request.FILES['profile_image']
